+++ src/gui/widgets/AudioFaderBox.cpp
@@ -307,7 +307,7 @@ void
 AudioFaderBox::setFont(QFont f)
 {
     m_synthButton->setFont(f);
-    for (int i = 0; i < m_plugins.size(); ++i) m_plugins[i]->setFont(f);
+    for (size_t i = 0; i < m_plugins.size(); ++i) m_plugins[i]->setFont(f);
     m_vuMeter->setFont(f);
     m_inputLabel->setFont(f);
     m_outputLabel->setFont(f);
