+++ src/sound/MappedEvent.h
@@ -194,7 +194,7 @@ public:
         WarningImpreciseTimerTryRTC = 11,
     } FailureCode;      
 
-    MappedEvent(): m_trackId(NO_TRACK),
+    MappedEvent(): m_trackId((int)NO_TRACK),
                    m_instrument(0),
                    m_type(MidiNote),
                    m_data1(0),
@@ -229,7 +229,7 @@ public:
                 const RealTime &absTime,
                 const RealTime &duration,
                 const RealTime &audioStartMarker):
-        m_trackId(NO_TRACK),
+        m_trackId((int)NO_TRACK),
         m_instrument(id),
         m_type(type),
         m_data1(pitch),
@@ -252,7 +252,7 @@ public:
                 const RealTime &eventTime,
                 const RealTime &duration,
                 const RealTime &audioStartMarker):
-         m_trackId(NO_TRACK),
+         m_trackId((int)NO_TRACK),
          m_instrument(id),
          m_type(Audio),
          m_data1(audioID % 256),
@@ -275,7 +275,7 @@ public:
                 MappedEventType type,
                 MidiByte data1,
                 MidiByte data2):
-         m_trackId(NO_TRACK),
+         m_trackId((int)NO_TRACK),
          m_instrument(id),
          m_type(type),
          m_data1(data1),
@@ -294,7 +294,7 @@ public:
     MappedEvent(InstrumentId id,
                 MappedEventType type,
                 MidiByte data1):
-        m_trackId(NO_TRACK),
+        m_trackId((int)NO_TRACK),
         m_instrument(id),
         m_type(type),
         m_data1(data1),
@@ -315,7 +315,7 @@ public:
     //
     MappedEvent(InstrumentId id,
                 MappedEventType type):
-        m_trackId(NO_TRACK),
+        m_trackId((int)NO_TRACK),
         m_instrument(id),
         m_type(type),
         m_data1(0),
